requirements.txt
httpx>=0.25
rich>=13.0
python-dotenv>=1.0
urllib3>=2.0
